Implement generateSymbolTable assembler helper
This is a particularly nice change from the old LC-3 code
in that it removes almost all of the mutation—
in particular, we now only mutate environment bindings, not values,
and the whole state machine is in a `reduce` instead of a mutative loop,
which makes the control flow much cleaner.

Disregarding functional principles, the actual code is much nicer, too:
we no longer have a ton of cases for different argument counts,
which makes the assembler both more architecturally flexible
and less brittle for exotic inputs under the current spec.

All thanks are to ES6.