Export `Constants.BATCH_MODE_LIMIT`
"What"—you cry—"how on earth did tests not catch this?"
Well, we were mocking out the value (to change it from `0x1000` to `4`)
so we didn't actually make sure it would work with the constant value.
Of course, when we ran it in the actual simulator,
we got `NaN`s because of adding `undefined` to things, etc.